create trip controller
